Refactor covid data widget
We would like to update data widget being shown on the covid data card
on the home screen. Currently a lot of the logic in the CovidData
context and feature is implicitly dependent on the format of the data
being pre calculated to a set of trends. In order to better support
changes in this feature, We would like to:

- More fully separate the view layer calculation of aggregates and the
raw data itself.

- More fully separate the parsing of network data and the applications
internal concept of a set of covid data.

This commit:
Is a refactor of the covid data analytics widget feature. No new
functionality has been added.