Introduce line chart for covid data
Why:
We would like to show a line chart in the covid data card to indicate to
users the trends of the virus.

This commit:

Introduces a CovidDataInfo component to render a the set of covid data
analytic info as a set of labeled visual elements.

Introduces a LineChart component which can render a scaled svg of a
provided data set as a curve.

Introduced a SvgPath module to handle the building of an svg path
command from a provided set of XYCoordinates.