Consolidate device logic
Why:

Currently we multiple places were device information is determined, we
would like to consolidate this logic. In particular, we have two contexts
that are modeling the same domain concept of device api access and
state, PermissionsContext and SystemServicesContext. We would like to
use a single concept within the app.

This commit:
Consolidates all of the device logic into a single module, Device Group
self assessment by feature and combines the logic from the
SystemServicesContext into the PermissionsContext. A future commit could
likely simplify the logic within.