Update symptom history date format
Why:
As a contact tracer speaking with a user, it is helpful to be able to
refer to the day of the week for a given symptom history entry. We would
like to show the day of week in addition to the calendar date on each
symptom history entry.

This commit:
Updates the symptom history date format and also opportunistically
organizes some of the components into sub features in the SymptomHistory
module.

Co-Authored-By: matt buckley <matt@nicethings.io>