Matomo Analytics Events (optional) (#625)
#### Why:
We'd like to implement a default set of analytics events to be tracked for users who have consented to providing anonymized data. 

#### This commit:
This commit introduces a small number of events that are sent to the Matomo analytics server for users who have opted in. In addition, this commit adjusts the preexisting event payload structure to send a category, action, and name.