Use covidactnow api (#631)
Why:
We would like to use the covidactnow.org api as it has a data set that
will allow us to show covid statistic and metrics for both states and
counties.

This commit:
Introduces an api fetch that consumes the covidactnow.org data. We also
introduced the `ts.data.json` package to use json decoders for decoding
the response json.

Note that the package name includes several `.` characters with metro
bundler does not handle correctly when calculating the sha1 hash for the
cache key in the bundler haste map. We resolved this issue by
introducing a postinstall script for yarn that will rename the package
within nodemodules.

Co-Authored-By: matt buckley <matt@nicethings.io>

Co-authored-by: matt buckley <matt@nicethings.io>