Fix various issues with product analytics
Why:
Currently we have a handful of minor visual and logic bugs in the
product analytics feature.

This commit:
- Verifies if the correct Matomo variables have been provided before
initializing the Matomo SDK
- moves the logic of determining if the build should enable Product
analytics out of the Product analytics context to the app initialization
to reduce complexity in the context.
- Separates the analytics consent screen into two components, one for
the onboarding activation flow and one for the settings screen to reduce
complexity in the navigation logic and reduce complexity in the forms
themselves.
- explicitly states if product analytics should be enabled rather than
inferring from the presence of the Matomo url
- removes the attempted tracking of a user opting into product analytics
as its current behavior only tracks when a user opts out of sharing
data, but still labels it opting in. We will add this back in a future
commit.
- removes the tracking call of if a user received an exposure
notification as this caused crashes in some cases. We will add this back
in a future commit.