Remove file processing limit for iOS 13.7+ (#642)
### Why
Devices running `iOS 13.7+` are not limited to processing a specific number of key archives in a given day. For this reason, we want to allow the app to process as many files as are available on a given day. 
<img width="823" alt="Screen Shot 2020-11-09 at 12 19 16 PM" src="https://user-images.githubusercontent.com/2637355/98574514-05035a80-2286-11eb-811b-5990a083bc1b.png">


### This Commit
This commit removes the upper limit on the number of files the iOS app processes per day for users running `iOS 13.7+`