Reset product analytics on delete all data
Why:
When a user elects to 'Delete all data' we would like to be deleting the
data related to remembering that the user has opted into sharing product
analytics.

This commit:
Adds the logic the to product analytics context to reset and invokes
this logic in the delete all data feature.