Update localizations
Why:
Currently we are not handling localizations for 'Back' navigations or
datetimes from dayjs

We would like to have dayjs to be localized correctly when the user
updates the locale and for the Back button to be translated.

This commit:
- Updates the navigation header left button to use a translated label.
- Updates the language.ts file to correctly determine the correct IETF
language tag for the current locale and sets the dayjs locale to that.
Note: that some languages the app supports, are not supported by dayjs,
e.g. hmn and so