Handle edge cases in en permissions activation
Why:

currently, there are various cases in which we do not present the
correct information to the user on how to enable Exposure Notifications.
These cases arise on the ActivateExposureNotifications and
ExposureDetectionStatus screens.

This commit:

Updates the exposureNotifications.request function in the
PermissionsContext to not only request authorization from the GAEN API,
but also return the authorization status determined from the request to
the GAEN API Adds a side effect to the ActivateExposureNotifications
screen to navigate to the next screen in the Activation flow if the
exposure notification status is enabled Adds an Alert to the
ActiveExposureNotifications screen that tells the user how to enable
exposure notifications if they request authorization from the GAEN API,
but need to change something in the app settings in order for exposure
notifications to be enabled Updates the copy for the Alert that explains
to the user how to enable exposure notifications in the app settings

Co-Authored-By: John Schoeman johnschoeman1617@gmail.com
Co-Authored-By: Matt Buckley matt@nicethings.io
Co-Authored-By: Juan Cruz <juancruzgs>