Remove gaen strategy abstraction
Why:
In previous iterations of this project there was a product need to be
able to support different contact tracing protocols, such as GAEN or
gps location redaction strategies. During that time an abstraction for
a contact tracing strategy was introduced. Given that this app is now
only targeting GAEN the strategy abstraction can be removed and the
related code can be simplified.

This commit:
removes the gaen strategy object and inlines its functionality.