Provide better error messages on request gaen api
Why:
There are many ways that the request to enable gean permissions for the
app could fail. Most of these response failures mean that the user needs to
manually enable the permissions for the app. We would like to be able to
provide specific instructions to the user on how to enable the
permissions.

this commit:
Refactors the api for the requestAuthorization function in the gaen
native module such that it is more clear what the native layer's
response actually was and to make it easier to respond to the specific
error. We provide a mapping in the native layer for a complete list of
error responses such that we can respond to each error case on an
individual basis.

Co-Authored-By: John Schoeman <johnschoeman1617@gmail.com>
Co-Authored-By; Matt Buckley <matt@nicethings.io>