Show notification status on home screen if off
Why: we would like to tell the user if their notifications are off.

This commit:
Adds a card to the home screen that informs the user if their
notifications are off. If the card is pressed and notification
permissions have not been previously denied, the app will request
notification permissions from the OS. If notifications have been
previously, the app will open the app settings.