Fetch custom copy from remote endpoint
Why:
We would like for the custom copy to be received from a remote cms and
cdn so that we may update the copy without needing to release a new
version of the app.

This commit:
Adds the networking logic to fetch the custom copy and adds it to the
useuseCustomCopy hook. We fallback to the static copy set at build time
if the network request fails for whatever reason.

Co-Authored-By: Matt Buckley <matt@nicethings.io>