Refactor open link logic
Why:
Currently production versions of the app will fail to open the
healthCheck url as we are using the `Linking.canOpenUrl` function which
checks against the LSApplicationQueriesSchemes value in the info.plist
on iOS which we do not have currently set. Given that we are only ever
needing to support https schemes this extra check is extraneous and can
be removed.

This commit:
Refactors the logic to open the health check link.