Reset locale on delete data
Why:
Currently when a user deletes their data we delete the saved locale
override token, but do not reset the in-app language setting in i18next.
This is confusing to the user as it seems like we didnt delete the
language token even though we did. We would like to reset the current
locale to the default of 'en' when the user deletes their data to avoid
this confusion.

This commit:
Adds a resetUserLocale function to the languages.ts module and calls it
the DeleteConfirmation component on deleteAllData.