Update icon used for the share link (#815)
Why:
Currently the icon used for the share link is named based off the
content. We would like for this image to be configurable, so the name
should be changed to be based on the usage rather than the content.

This commit:
Replaces `HugEmoji.png` with `ShareImage.png`, which is the standard
material UI 'share' graphic. We also remove some unused iconography
styles.