Update quarantine end date on exposure history
Why:
Currently we have a bug where we incorrectly calculate the end date of
a recommended quarantine for users that have received an exposure
notification. Additionally quarantines that have been completed will
continue to show a '0' remaining days and an end date of the current
day.

This commit:
Determines the last day of quarantine by adding quarantine days
remaining to the current day instead of the date of the exposure itself
and checks if the quarantine is over.