Handle dont share action in escrow verification (#833)
Why:
During the submit your phone number step of the Escrow Verification
flow, an OS level prompt asks if the user would like to share their
devices Random IDs. If the user chooses 'Dont Share' the os will raise
an error which we are not handling.

This commit:
- Updates the error handling for the submitPhoneNumber and submitKeys
iOS native method to return ENDomainErrors when the os returns them.

- Updates the javascript side error handling for the submitPhoneNumber
and submitKeys native method. Note that the api for the native methods
can return both network error and ENDomainErrors which do not follow the
same error format. We introduced a `handleNetworkErrror()` and
`handleENDomainError()` to handle these two cases. A future refactor
might consider updating the native method to better separate the error
conditions.