GAEN-743 Analytic Event For Sending Notification (#894)
* GAEN-743 Analytic Event For Sending Notification

Create a StateUpdatedWorker class in Debug and the Release package.
Use an intent inside of the DebugMenuModule to actually trigger the ExposureNotificationBroadcastReceiver.class
Create a new ExposureResult to be returned by the refreshWithDailySummaries

* GAEN-743 Notification Analytics

Add an intent param for the runOnce function in the StateUpdateWorker. Only the debug class will actually utilize it. Release has the same param but ignores it.
Removes the DeviceEventManagerModule.RCTDeviceEventEmitter from the Worker code.

* GAEN-743 Updates for styling